acme: add keyboard shortcuts
Add keybaord shortcuts for common actions using the command keys, in a
new key table outside of the standard text processing flow, so that
operations on a window (like Del) can succeed.

The command keys are:
    %S          Put
    %-shift-S   Putall
    %-shift-F   Look
    %W          Del
    %-shift-W   Delcol
    %N          New
    %-shift-N   Newcol
    %;          Spell

To support this, a new executerunestr function is added which could be
utlized to implement an exec or eval ctl command.