acme: tab-aware up/down movement
The textbacklinewidthtext and forwardlinewidth functions have been
extended to optionally expand \t into tabstop in the returned width,
and the Kup and Kdown handlers have been updated to decrement the
offset width by tabstop when encountering a \t. This enables up/down
to move in a vertical line even if the line above or below has more
or less indentation.